requirements.txt
faiss-cpu==1.7.4
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
openai==1.6.1
pydantic-settings==2.2.1
pypdf2==3.0.1
python-dotenv
streamlit==1.35.0
tiktoken==0.5.2
langchain_groq


